Moved Nuklear rendering code to a separate NkRenderer class

+++ CoreEngine/src/jx3d/graphics/IndexBuffer.java
@@ -3,6 +3,7 @@ package jx3d.graphics;
 import jx3d.graphics.opengl.GLGraphics;
 import jx3d.graphics.opengl.GLIndexBuffer;
 
+import java.nio.ByteBuffer;
 import java.nio.ShortBuffer;
 
 /**
@@ -92,7 +93,7 @@ public abstract class IndexBuffer extends Buffer {
     /**
      * Map this buffer's data into the client's address space.
      */
-    public abstract ShortBuffer map();
+    public abstract ByteBuffer map();
 
     /**
      * Unmaps the buffer.<br>
